requirements.txt
django
gunicorn
django-heroku
paho-mqtt==1.3.0
requests==2.20.1
